[python-frontend] Clean code
